chore: rename, move and create test code/test cases, with other awkard small changes
