+++ application/extension/test/dr/dr_19_let_touyoushi_large0.json
@@ -578,7 +578,7 @@
                       "majors": [
                         "HUM"
                       ],
-                      "like": "[中国語|英語|古典語|ドイツ語|朝鮮語|アラビア語]"
+                      "like": "中国語|英語|古典語|ドイツ語|朝鮮語|アラビア語"
                     }
                   }
                 }
